+++ changelog.md
@@ -1,2 +1,5 @@
+### Entangled 1.3.21
+- Added Japanese translations (thanks to Twister716!)
+
 ### Entangled 1.3.20
 - Initial release of Entangled for Minecraft 1.21.11